requirements.txt
djangorestframework==3.10.3
djangorestframework-jwt
django-cors-headers==3.1.1
